<FEAT>(MEEPMEEP/AUTO): added 45 deg spline implementation
